revert PolicyPrediction (confidence, priority is tuple); commented out validate_against_domain which needs to be adapted in rulepolicy
