+++ rasa/core/policies/memoization.py
@@ -25,7 +25,7 @@ from rasa.core.featurizers.tracker_featurizers import (
 )
 from rasa.core.featurizers.tracker_featurizers import FEATURIZER_FILE
 from rasa.shared.exceptions import FileIOException
-from rasa.core.policies.policy import PolicyPrediction2, PolicyGraphComponent
+from rasa.core.policies.policy import PolicyPrediction, PolicyGraphComponent
 from rasa.shared.core.trackers import DialogueStateTracker
 from rasa.shared.core.generator import TrackerWithCachedStates
 from rasa.shared.utils.io import is_logging_disabled
@@ -228,7 +228,7 @@ class MemoizationPolicyGraphComponent(PolicyGraphComponent):
         domain: Domain,
         precomputations: Optional[MessageContainerForCoreFeaturization] = None,
         **kwargs: Any,
-    ) -> PolicyPrediction2:
+    ) -> PolicyPrediction:
         """Predicts the next action the bot should take after seeing the tracker.
 
         Args:
