+++ rasa/core/policies/unexpected_intent_policy.py
@@ -33,7 +33,7 @@ from rasa.core.featurizers.single_state_featurizer import (
 )
 from rasa.shared.core.generator import TrackerWithCachedStates
 from rasa.core.constants import DIALOGUE, POLICY_MAX_HISTORY
-from rasa.core.policies.policy import PolicyPrediction2
+from rasa.core.policies.policy import PolicyPrediction
 from rasa.core.policies.ted_policy import (
     LABEL_KEY,
     LABEL_SUB_KEY,
@@ -564,7 +564,7 @@ class UnexpecTEDIntentPolicyGraphComponent(TEDPolicy):
         domain: Domain,
         precomputations: Optional[MessageContainerForCoreFeaturization],
         **kwargs: Any,
-    ) -> PolicyPrediction2:
+    ) -> PolicyPrediction:
         """Predicts the next action the bot should take after seeing the tracker.
 
         Args:
