+++ tests/core/test_ensemble.py
@@ -6,7 +6,7 @@ from typing import List, Tuple
 from rasa.engine.graph import ExecutionContext
 from rasa.engine.storage.resource import Resource
 from rasa.engine.storage.storage import ModelStorage
-from rasa.core.policies.policy import PolicyPrediction2
+from rasa.core.policies.policy import PolicyPrediction
 from rasa.core.policies.ensemble import (
     DefaultPolicyPredictionEnsemble,
     PolicyPredictionEnsemble,
@@ -21,14 +21,6 @@ from rasa.shared.core.constants import ACTION_LISTEN_NAME
 def test_warn_warn_rule_policy_not_contained():
     with pytest.warns(UserWarning, match="'RulePolicy' is not included"):
         PolicyPredictionEnsemble.warn_rule_policy_not_contained(ensemble=[])
-
-
-def test_assert_compatibility_with_domain():
-    domain = Domain.load("data/test_domains/form.yml")
-    with pytest.raises(InvalidDomain, match="You have defined a form action, but"):
-        PolicyPredictionEnsemble.assert_compatibility_with_domain(
-            ensemble=[], domain=domain
-        )
 
 
 def test_warn_priorities_not_unique():
@@ -54,7 +46,7 @@ def test_ensemble_predict(default_ensemble: DefaultPolicyPredictionEnsemble):
     tracker = DialogueStateTracker.from_events("test", [UserUttered("hi")], [])
     num_actions = len(domain.action_names_or_texts)
     predictions = [
-        PolicyPrediction2(
+        PolicyPrediction(
             policy_name=str(idx), probabilities=[idx] * num_actions, policy_priority=idx
         )
         for idx in range(2)
@@ -80,7 +72,7 @@ def test_default_predict_excludes_rejected_action(
     )
     num_actions = len(domain.action_names_or_texts)
     predictions = [
-        PolicyPrediction2(
+        PolicyPrediction(
             policy_name=str(idx), probabilities=[1.0] * num_actions, policy_priority=idx
         )
         for idx in range(2)
@@ -99,7 +91,7 @@ def test_default_predict_excludes_rejected_action(
             (
                 # highest probability and highest priority
                 [
-                    PolicyPrediction2(
+                    PolicyPrediction(
                         policy_name=str(idx),
                         probabilities=[idx] * 3,
                         policy_priority=idx,
@@ -111,7 +103,7 @@ def test_default_predict_excludes_rejected_action(
             (
                 # highest probability wins even if priority is low
                 [
-                    PolicyPrediction2(
+                    PolicyPrediction(
                         policy_name=str(idx),
                         probabilities=[idx] * 3,
                         policy_priority=idx,
@@ -123,13 +115,13 @@ def test_default_predict_excludes_rejected_action(
             (
                 # "end to end" prediction supersedes others
                 [
-                    PolicyPrediction2(
+                    PolicyPrediction(
                         policy_name="policy using user text but max prob 0.0 wins",
                         probabilities=[0.0],
                         policy_priority=0,
                         is_end_to_end_prediction=True,
                     ),
-                    PolicyPrediction2(
+                    PolicyPrediction(
                         policy_name="policy not using user text but max prob 1.0",
                         probabilities=[1.0],
                         policy_priority=1,
@@ -141,19 +133,19 @@ def test_default_predict_excludes_rejected_action(
             (
                 # "no user" prediction supsersedes even the end to end ones
                 [
-                    PolicyPrediction2(
+                    PolicyPrediction(
                         policy_name="'no user' with smallest max. prob",
                         probabilities=[0.0],
                         policy_priority=0,
                         is_no_user_prediction=True,
                     ),
-                    PolicyPrediction2(
+                    PolicyPrediction(
                         policy_name="'end2end' with higher prob and priority",
                         probabilities=[1.0],
                         policy_priority=1,
                         is_end_to_end_prediction=True,
                     ),
-                    PolicyPrediction2(
+                    PolicyPrediction(
                         policy_name="highest prob and highest priority",
                         probabilities=[2.0],
                         policy_priority=2,
@@ -167,7 +159,7 @@ def test_default_predict_excludes_rejected_action(
 )
 def test_default_combine_predictions(
     default_ensemble: DefaultPolicyPredictionEnsemble,
-    predictions_and_expected_winner_idx: Tuple[List[PolicyPrediction2], int],
+    predictions_and_expected_winner_idx: Tuple[List[PolicyPrediction], int],
     last_action_was_action_listen: bool,
 ):
     predictions, expected_winner_idx = predictions_and_expected_winner_idx
